+++ IPB/LoadGameScreen.cpp
@@ -38,7 +38,7 @@ void LoadGameScreen::render()
 		backBtn->render(renderer);
 	}
 
-	ScreenTitle->setPosition(ScreenTitle->getX(), y);
+	ScreenTitle->setPosition(ScreenTitle->getX(), static_cast<int>(y));
 }
