+++ IPB/PauseScreen.cpp
@@ -37,7 +37,7 @@ void PauseScreen::render()
 		quitGameBtn->render(renderer);
 	}
 
-	gameTitle->setPosition(gameTitle->getX(), y);
+	gameTitle->setPosition(gameTitle->getX(), static_cast<int>(y));
 }
