+++ IPB/SelectLevelScreen.cpp
@@ -42,7 +42,7 @@ void SelectLevelScreen::render()
 		BackBtn->render(renderer);
 	}
 
-	ScreenTitle->setPosition(ScreenTitle->getX(), y);
+	ScreenTitle->setPosition(ScreenTitle->getX(), static_cast<int>(y));
 }
