+++ IPB/Word.cpp
@@ -7,7 +7,7 @@
 Word::Word(string str, LTexture * imagesSpriteSheet, int x, int y, float scale)
 {
 	this->scale = scale;
-	this->charSpacing = 30 * scale;
+	this->charSpacing = static_cast<int>(30 * scale);
 	this->txtTexture = imagesSpriteSheet;
 	this->setText(str);
 	this->setPosition(x, y);
@@ -21,13 +21,13 @@ int Word::getTextLength()
 }
 void Word::setXCentered()
 {
-	int x = constants::WINDOW_WIDTH - renderWord.length() * constants::CHARACTER_WIDTH * scale + (renderWord.length() - 1) * 30 * scale;
+	int x = static_cast<int>(constants::WINDOW_WIDTH - renderWord.length() * constants::CHARACTER_WIDTH * scale + (renderWord.length() - 1) * 30 * scale);
 	x = x / 2;
 	setPosition(x, position.y);
 }
 void Word::setYCentered()
 {
-	int y = constants::WINDOW_HEIGHT - constants::CHARACTER_HEIGHT * scale;
+	int y = static_cast<int>(constants::WINDOW_HEIGHT - constants::CHARACTER_HEIGHT * scale);
 	y = y / 2;
 	setPosition(position.x, y);
 }
@@ -35,14 +35,14 @@ void Word::setPosition(int x, int y)
 {
 	this->position.x = x;
 	this->position.y = y;
-	for (int i = 0; i < this->renderWord.length(); i++)
+	for (int i = 0; i < static_cast<int>(this->renderWord.length()); i++)
 	{
 		characters[i].setPosition(x + i * charSpacing, y);
 	}
 }
 void Word::render(SDL_Renderer * gRenderer)
 {
-	for (int i = 0; i < this->renderWord.length(); i++)
+	for (int i = 0; i < static_cast<int>(this->renderWord.length()); i++)
 	{
 		characters[i].render(gRenderer);
 	}
@@ -55,7 +55,7 @@ void Word::setText(string str)
 		delete[] this->characters;
 	}
 	this->characters = new Character[str.length()];
-	for (int i = 0; i < this->renderWord.length(); i++)
+	for (int i = 0; i < static_cast<int>(this->renderWord.length()); i++)
 	{
 		this->characters[i].setTexture(this->txtTexture);
 		this->characters[i].setChar(str[i]);
